refactor: move more files
