+++ packages/consumption/src/modules/attributes/local/attributeTypes/ThirdPartyRelationshipAttribute.ts
@@ -3,7 +3,7 @@ import { IRelationshipAttribute, RelationshipAttribute, RelationshipAttributeJSO
 import { CoreAddress, CoreId, ICoreAddress, ICoreId } from "@nmshd/core-types";
 import { nameof } from "ts-simple-nameof";
 import { ConsumptionCoreErrors } from "../../../../consumption/ConsumptionCoreErrors";
-import { IReceivedAttributeDeletionInfo, ReceivedAttributeDeletionInfo, ReceivedAttributeDeletionInfoJSON, ReceivedAttributeDeletionStatus } from "../sharingDetails";
+import { IReceivedAttributeDeletionInfo, ReceivedAttributeDeletionInfo, ReceivedAttributeDeletionInfoJSON, ReceivedAttributeDeletionStatus } from "../deletionInfos";
 import { ILocalAttribute, LocalAttribute, LocalAttributeJSON } from "./LocalAttribute";
 
 export interface ThirdPartyRelationshipAttributeJSON extends LocalAttributeJSON {
