+++ packages/consumption/src/modules/attributes/local/deletionInfos/AbstractAttributeDeletionInfo.ts
