+++ packages/consumption/src/modules/attributes/local/deletionInfos/EmittedAttributeDeletionInfo.ts
