+++ packages/consumption/src/modules/attributes/local/deletionInfos/ReceivedAttributeDeletionInfo.ts
