+++ packages/consumption/src/modules/attributes/local/deletionInfos/index.ts
