+++ packages/consumption/src/modules/attributes/local/sharingDetails/index.ts
@@ -1,2 +1,1 @@
-export * from "./deletionInfos";
 export * from "./ForwardedSharingDetails";
